[move]SlotItemTypeクラスをlogbook.constantsパッケージへ移動
