+++ src/main/java/logbook/bean/Ship.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 import javax.json.JsonObject;
 
 import logbook.constants.ShipType;
+import logbook.constants.SlotItemType;
 import logbook.internal.JsonHelper;
 import logbook.internal.Ships;
-import logbook.internal.SlotItemType;
 import lombok.Data;
 
 /**
