+++ src/main/java/logbook/internal/Ships.java
@@ -43,6 +43,7 @@ import logbook.bean.Stype;
 import logbook.bean.StypeCollection;
 import logbook.common.Messages;
 import logbook.constants.ShipType;
+import logbook.constants.SlotItemType;
 import logbook.internal.Tuple.Pair;
 import logbook.plugin.PluginServices;
 import lombok.AllArgsConstructor;
